+++ addOns/selenium/src/main/resources/org/zaproxy/zap/extension/selenium/resources/stealth/navigator.hardwareconcurrency.js
@@ -1,0 +1,6 @@
+// "4" is an arbitrary value. Using a random value may help evade detection.
+utils.replaceGetterWithProxy(
+    Object.getPrototypeOf(navigator),
+    'hardwareConcurrency',
+    utils.makeHandler().getterValue(4)
+)