+++ addOns/selenium/src/main/resources/org/zaproxy/zap/extension/selenium/resources/stealth/navigator.webdriver.js
@@ -1,0 +1,8 @@
+if (navigator.webdriver === false) {
+    // Post Chrome 89.0.4339.0 and already good
+} else if (navigator.webdriver === undefined) {
+    // Pre Chrome 89.0.4339.0 and already good
+} else {
+    // Pre Chrome 88.0.4291.0 and needs patching
+    delete Object.getPrototypeOf(navigator).webdriver
+}