+++ addOns/selenium/src/main/resources/org/zaproxy/zap/extension/selenium/resources/stealth/window.outerdimensions.js
@@ -1,0 +1,8 @@
+try {
+    if (!(window.outerWidth && window.outerHeight)) {
+        const windowFrame = 85 // probably OS and WM dependent
+        window.outerWidth = window.innerWidth
+        window.outerHeight = window.innerHeight + windowFrame
+    }
+} catch (err) {
+}