+++ addOns/selenium/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+- Added stealth mode to make the browser appear to be controlled by a user (Chrome only).
 
 ## [15.40.0] - 2025-09-02
 ### Changed